# Startup `flutter` faster (use app-jit snapshot)