@christopherfujino 2022-09-15
> I think I had tried this a few years ago, but I found that the increased compilation time was really quite noticeable. Has performance of app-jit generation improved in recent years?
> 
> I definitely don't mind trying this out though

Ahh ok, I remember you tried this but I forgot why it got reverted. Tool compilation should only be an issue either during flutter upgrade, or for the first time after a user checks out a new commit. Is this not reasonable?