@christopherfujino 2022-09-15
Maybe the answer is we should also benchmark tool compilation?