@christopherfujino 2022-09-15
> I don't really think that makes sense to track, unless we want to also track the time it takes to download all of the binaries. It only needs to be tolerable for folks that regularly contribute to the tool

I didn't mean adding a skiaperf benchmark, but at least compiling it both ways n times and documenting on this PR