@jensjoha 2022-09-16
I did that (on my Linux machine):

Kernel and app-jit is done basically as-is (i.e. `--snapshot=` and `--snapshot-kind=` flags to dart), whereas exe is done via `dart compile exe`.

Runtimes for snapshotting as measured via `time`.

  | kernel | app-jit | exe
-- | -- | -- | --
  | 10.035 | 10.44 | 32.967
  | 9.622 | 10.302 | 33.132
  | 9.435 | 10.253 | 33.231
  | 9.462 | 10.287 | 32.831
  | 9.49 | 10.186 | 32.535
  |   |   |  
Average | 9.6088 | 10.2936 | 32.9392
  | 100.00% | 107.13% | 342.80%