@jonahwilliams 2022-09-16
Lets wait until we get the startup benchmarks running: https://github.com/flutter/flutter/issues/111748