@zanderso 2022-09-29
> I think I had tried this a few years ago, but I found that the increased compilation time was really quite noticeable. Has performance of app-jit generation improved in recent years?

Previously: https://github.com/flutter/flutter/issues/30102. It was reverted due to a crash in the VM, and it looks like we didn't follow up =(. Presumably in the intervening years, the crasher has been fixed.