@jensjoha 2022-09-30
No. This failed to. I fixed it in https://github.com/flutter/flutter/pull/111879.